REFACTOR move pollIncidents onto its own file
